chore: cleanup dir and add typedoc
