+++ client/packages/cli/config/transfer.ts
@@ -14,7 +14,7 @@ export default {
             to: "5EoHBHDBNj61SbqNPcgYzwHXY1xAroduRP3M99iSMZ8kwvgp",
             amount: "0.05", // in ROC
             insurance: "0.8", // in TRN
-            reward: "0.0001", // in TRN
+            reward: "4", // in TRN
         },
     ],
     sequential: false,
