+++ client/packages/executor/package.json
@@ -4,7 +4,7 @@
   "main": "index.js",
   "license": "MIT",
   "scripts": {
-    "start": "DEBUG=instance-manager,execution-manager,substrate-relayer,circuit-relayer,circuit-listener ts-node ./index.ts",
+    "start": "DEBUG=instance-manager,execution-manager,substrate-relayer,circuit-relayer,circuit-listener ts-node ./src/index.ts",
     "fmt": "npx prettier --no-semi --arrow-parens avoid --print-width 80 --write ./*.ts ./**/*.ts ./**/**/*.ts"
   },
   "dependencies": {
@@ -18,6 +18,7 @@
     "dotenv": "16.0.0",
     "rxjs": "^7.5.7",
     "ts-node": "10.7.0",
+    "typedoc": "^0.23.21",
     "typescript": "^4.8.4"
   }
 }
