+++ client/packages/executor/src/strategy/index.ts
