+++ client/packages/executor/typedoc.json
@@ -1,0 +1,7 @@
+{
+    "entryPoints": ["src/index.ts"],
+    "out": "docs",
+    "tsconfig": "./tsconfig.json",
+    "includes": ["src"],
+    "excludeNotDocumented": false
+}
